Read simulation parameters from yaml file
This is some very basic functionality to read the parameters for a
simulation from an input file, which moves towards every directory being
a complete record of the experiment that took place.

This is also begins to make the scripts to submit jobs to a scheduler
more flexible, allowing for a single script to work for a number of
simulation types, and working towards a single script for all
simulations with the parameters defined in the yaml file.