Update run_simulation to modify molecule
This update includes a very basic capability to modify the molecule that
is used in the simulation. This is done by specifying the molecule name,
distance, radius, and moment of inertia. Currently this is just a list
of the command line options, although as I develop the capability
further I will be able to make this more general.